first version working

+++ src/transformers/generation_utils.py
@@ -402,7 +402,7 @@ class GenerationMixin:
     ) -> torch.LongTensor:
         decoder_start_token_id = self._get_decoder_start_token_id(decoder_start_token_id, bos_token_id)
         decoder_input_ids = (
-            torch.ones((input_ids.shape[0], 1), dtype=input_ids.dtype, device=input_ids.device)
+            torch.ones((input_ids.shape[0], 1), dtype=torch.long, device=input_ids.device)
             * decoder_start_token_id
         )
         return decoder_input_ids
